move SyncRequest and related types into separate module
